requirements.txt
Flask==3.0.2
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27
Werkzeug==3.0.1